fetching all procedures
